Upgrades Scribe to PartiQL 0.14.5
