+++ build.gradle.kts
@@ -20,7 +20,7 @@ object Versions {
     const val jline = "3.21.0"
     const val junit5 = "5.9.3"
     const val picoCli = "4.7.0"
-    const val partiql = "0.14.4"
+    const val partiql = "0.14.5"
 }
 
 object Deps {
