+++ src/main/kotlin/org/partiql/scribe/sql/rewriters/Rewriter.kt
@@ -1,0 +1,9 @@
+package org.partiql.scribe.sql.rewriters
+
+import org.partiql.ast.AstNode
+
+internal interface Rewriter {
+
+    fun apply(ast: AstNode): AstNode
+
+}
